Login Authentication Not Working
Previously, the login feature was not fully implemented and
connected to the database. Now, users can log in and will be
redirected to their respective pages based on their credentials.
Below is a list of other notable changes, since this is a rather
large commit:

- All content pages (Administrator, Member, Instructor) have all been
moved into their own folders with their own `Web.Config` files for
user authentication

- GridViews for Member and Instructor pages display the proper information
based on the user that is logged in

- Fixed the LINQ queries for both pages to remove unnecessary columns
displayed in the final GridView

- Moved the database Connection string declaration to Web.Config

- Going forward, the Connection string should be changed there instead of
in every spot in the file and not uploaded as changes when commits are made.
There is a chance that the string is now agnostic and won't need to be changed,
but I could be wrong

- Cleaned up some methodology in Administrator.aspx.cs where some variables
were serving as intermediary but otherwise pointless

The last step is to finish implementation for administrator features.