Implement Delete Member/Instructor and Add Member to Section
Deleting Members and Instructors now works. Also: members can be added
to sections based on the Member and Instructor IDs selected from the
drop down lists, along with the selected value of the section drop down
list. Additionally, this commit contains some changes to the database
to add ON DELETE CASCADE to a few of the tables.

4 SQL files have also been added for easy data testing when modifying
or deleting records. After closing the web app, you can click the script
button on the table interface and run any one of these scripts to
reload the data into the proper table.